refactor: extract project activity issue and action routes
